Fix FriendsApi, Add friendsModel
